requirements.txt
robotframework
robotframework-seleniumlibrary
robotframework-datadriver
opencv-python
visual-comparison


